# Preallocate space for the GetPieces result vector